crawler/requirements.txt
requests
beautifulsoup4
lxml
curl_cffi
psycopg2-binary
python-slugify
